+++ mobile/ViewMediaTableViewController.m
@@ -127,7 +127,7 @@
     self.media.commentsCount++;
     
     // Update the likes in the interface.
-    [[[[[self.data objectAtIndex:0] objectAtIndex:0] objectForKey:@"Stats"] objectAtIndex:2] setObject:[NSString stringWithFormat:@"%d", self.media.commentsCount] forKey:@"count"];
+    [[[[[self.data objectAtIndex:0] objectAtIndex:0] objectForKey:@"Stats"] objectAtIndex:2] setObject:[NSString stringWithFormat:@"%ld", (long)self.media.commentsCount] forKey:@"count"];
     
     // Update the comments count in the interface.
     [[[self.data objectAtIndex:2] objectAtIndex:0] setObject:[NSString stringWithFormat:@"عرض التعليقات الـ %ld أو إضافة", (long)self.media.commentsCount] forKey:@"Add"];
@@ -268,7 +268,7 @@
                 self.media.likesCount++;
                 
                 // Update the likes in the interface.
-                [[[[[self.data objectAtIndex:0] objectAtIndex:0] objectForKey:@"Stats"] objectAtIndex:1] setObject:[NSString stringWithFormat:@"%d", self.media.likesCount] forKey:@"count"];
+                [[[[[self.data objectAtIndex:0] objectAtIndex:0] objectForKey:@"Stats"] objectAtIndex:1] setObject:[NSString stringWithFormat:@"%ld", (long)self.media.likesCount] forKey:@"count"];
                 
                 [self.likeButton setEnabled:NO];
